Remove sphinx gallery from sphinx doc build.

+++ docs/build_doc.py
@@ -1,31 +1,18 @@
-import argparse, subprocess, platform, os, shutil
-
-def open_url(url):
-    ''' Utility function '''
-    system_name = platform.system()
-    if system_name == 'Darwin':
-        browser = 'open'
-    elif system_name == 'Linux':
-        browser = 'xdg-open'
-    elif system_name == 'Windows':
-        browser = 'explorer.exe'
-    else:
-        return
-    subprocess.call([browser, url])
+import argparse, subprocess, platform, os, shutil, webbrowser
 
 def clean():
     files = []
     folders = ['_build', 'doxygen/html', 'doxygen/latex', 'doxygen/xml']
     for f in files:
-        os.remove(f)
+        if os.path.isfile(f):
+            os.remove(f)
 
     for f in folders:
-        shutil.rmtree(f)
+        if os.path.isdir(f):
+            shutil.rmtree(f)
 
 def main():
     parser = argparse.ArgumentParser()
-    parser.add_argument('--build_tutorial', action='store_true',
-        default=False, help='Whether we should run the ipython notebook')
     parser.add_argument('--build_doxygen', action='store_true',
         default=False)
     parser.add_argument('--rtd', action='store_true',
@@ -34,7 +21,6 @@ def main():
         default=False, help='Remove build artifacts')
 
     args = parser.parse_args()
-    is_build_tutorial = args.build_tutorial
     is_build_doxygen = args.build_doxygen
     is_on_rtd = args.rtd
     is_clean = args.clean
@@ -51,9 +37,6 @@ def main():
             print(e)
 
     env = dict(os.environ)
-    # print(env)
-    if is_build_tutorial:
-        env['UNREALCV_BUILD_TUTORIAL'] = 'True'
 
     if is_on_rtd:
         env['READTHEDOCS'] = 'True'
@@ -62,14 +45,14 @@ def main():
         'sphinx-build', '-n',
         '-b', 'html', # build format
         '.', '_build/html', # input, output folder
+        '-a',
     ]
     print(cmd)
     subprocess.call(cmd, env = env)
     # subprocess.call(cmd, env = os.environ)
 
     index_file = os.path.join('_build', 'html', 'index.html')
-    open_url(index_file)
-
+    webbrowser.open_new(index_file)
 
 if __name__ == '__main__':
     main()
